+++ src/main/java/com/yuuko/core/Cache.java
@@ -4,7 +4,7 @@ import com.google.api.services.youtube.model.SearchResult;
 import com.yuuko.core.commands.Command;
 import com.yuuko.core.commands.Module;
 import com.yuuko.core.commands.audio.handlers.AudioManagerManager;
-import net.dv8tion.jda.core.JDA;
+import com.yuuko.core.commands.audio.handlers.LavalinkManager;
 import net.dv8tion.jda.core.entities.User;
 import org.discordbots.api.client.DiscordBotListAPI;
 
@@ -12,7 +12,8 @@ import java.util.HashMap;
 import java.util.List;
 
 public class Cache {
-    public static JDA JDA;
+    public static net.dv8tion.jda.bot.sharding.ShardManager SHARD_MANAGER;
+    public static LavalinkManager LAVALINK;
     public static User BOT;
     public static List<Command> COMMANDS;
     public static List<Module> MODULES;