+++ src/main/java/com/yuuko/core/commands/audio/handlers/GuildAudioManager.java
@@ -1,30 +1,23 @@
 package com.yuuko.core.commands.audio.handlers;
 
-import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
-import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
+import com.yuuko.core.Cache;
+import lavalink.client.player.IPlayer;
 
 public class GuildAudioManager {
 
-    public final AudioPlayer player;
+    public final IPlayer player;
     public final TrackScheduler scheduler;
 
     /**
      * Creates a player and a track scheduler.
      * Without this class, trying to get persistent players, schedules, ect,
      * over an instanced module based bot would be pretty difficult.
-     * @param manager Audio player manager to use for creating the player.
+     * @param guildId Id of the guild to create a player for.
      */
-    public GuildAudioManager(AudioPlayerManager manager) {
-        player = manager.createPlayer();
+    public GuildAudioManager(String guildId) {
+        player = Cache.LAVALINK.createPlayer(guildId);
         scheduler = new TrackScheduler(player);
         player.addListener(scheduler);
-    }
-
-    /**
-     * @return Wrapper around AudioPlayer to use it as an AudioSendHandler.
-     */
-    public AudioPlayerSendHandler getSendHandler() {
-        return new AudioPlayerSendHandler(player);
     }
 
 }